finished the functionality of the start menu
finished the functionality of the start menu, but i still need to edit it to fit the needs of this game specifically. i have also started making a pause screen which would have the help area which would include the controls, and there would also be the option to start/end the game, stats and a log of all of the items you have discovered and lore/info on the items including the stats and where it originated from and what it does to certain weapons.