Finalised the Start Menu and Created the Pause Menu
Finalised the Start Menu and Created the Pause Menu. I still have to think of an idea for how the pause menu should look like but the start menu is mostly done, I haven't implemented the volume logic though and I haven't coded the different difficulties yet but I will get to that later in the game production phase. I am now going to create the death animation and random fact and I am going to implement the control and item uis. I actually locked in lol.