I locked in
Introduced the gamemanager, used for managing different scenes and different states of the game. Created a new scene for exclusively the managers and made them all singletons to decrease the amount of bugs we can get and make the code structurally more sound.

The game manager will also handle what you need to save. I also added the beginning of the save file, but right now it is not working. Will work on that next. I also need to work on fully integrating every file to work with this new game manager.