Completed LevelManager
I completely made everything compatible with gamemanager and created a levelmanager where I can put all of the functions that the buttons will do into that one file. It is not a singleton so it can be in every scene and it will improve readability immensely. Now everything works smoothly. The next thing I need to work on is the save load data stuff.