added health too
i added the health save data too, but there is this really weird glitch that basically reverts the save data, or maybe doesn't get it in the first place, even though every log says otherwise. I am leaving it with the bug because I suspect the bug is happening because we are constantly decreasing the value of the playerhealth, and somehow that reverts it to what it was before.