Saving
Created the npcinteract class and have added the mechanic to the pick_up_mechanic script. however, you need to press 'g' to interact with an npc but I want both items to be picked up and npc's to be interactable with 'f' so i am trying to merge them now.