Damage is possible working
I think i have got the jabsword function to work but i need a way to visualise what its hitting because i have coded the direction into it but i am unsure if it is following that code. but before that i need to sort out a solution to the cooldown because currently after an ability is used it can never be used again, and the ui is pretty buggy right now.