Got the jab function to work with a caveat
Since we still don't have a model for the player and player is still a sphere, the direction that the jab is hitting is going haywire, and only hits about 50% of the time. I could fix this by making the vectors 2d and changing some other variables, but I think it is better to wait for the player model to be ready and then I will further playtest. Also the gizmo is not working :(. Now I have to fix the death function and make the monster dissapear when dead.