Create mesh collider for main character
For some reason the mesh collider didn't work and was really unstable so I decided to use another technique where you just combine a bunch of primitive colliders so fit the mold of the character. It takes a bit of time but hopefully it will be quicker now because I have done it once and have the building blocks in place. Going to work on a satisfying and reponsive character movement script now.